requirements.txt
Django>=3.1
djangorestframework>=3.12
aspectlib>=1.5
coverage>=5.5
django-fake-model>=0.1.4
django-extensions>=3.1